money-mentor-backend/requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
dj-rest-auth
Django
django-allauth
django-ckeditor
django-cors-headers
django-filter
django-js-asset
djangocms-admin-style
djangorestframework
idna
importlib-metadata
Markdown
mysqlclient
oauthlib
Pillow
pycparser
PyJWT
python3-openid
pytz
requests
requests-oauthlib
sqlparse
urllib3
zipp
